Implemented Filtration
Implemented the filtration of the dataframe. Note, due to how it's made, technically one should only have to add the appropriate columns to the jobs dataframe and the filter matrix in order to condition against a new class (wherein how the new class should operate would just be defined in evalClass(), adding to the if-else).